+++ state_manager/routes/base.py
@@ -24,7 +24,7 @@ class BaseRouter(ABC):
         event_type: str,
         handler: Callable,
         *,
-        state_name: Optional[Union[str, list, tuple]] = None,
+        state_name: Optional[Union[str, List[Optional[str]]]] = None,
         filters: Optional[Tuple[Callable[..., bool], ...]] = None,
     ) -> None:
         if isinstance(state_name, str) or state_name is None:
